config/requirements.txt
pip==24.1.1
python-dotenv
Flask==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
itsdangerous==2.2.0
jsonify==0.5
mysql-connector-python==9.0.0
pyodbc==5.1.0
